remove functions

+++ translate/src/v3_delete_glossary.php
@@ -23,50 +23,34 @@
 //   title: Delete Glossary
 //   description: Delete Glossary
 //   usage: php v3_delete_glossary.php [--project_id "[Google Cloud Project ID]"] [--glossary_id "[Glossary ID]"]
-// [START translate_v3_delete_glossary]
 require_once __DIR__ . '/../vendor/autoload.php';
 
+if (count($argv) < 3 || count($argv) > 3) {
+    return printf("Usage: php %s PROJECT_ID GLOSSARY_ID\n", __FILE__);
+}
+list($_, $projectId, $glossaryId) = $argv;
+
+// [START translate_v3_delete_glossary]
 use Google\Cloud\Translate\V3\TranslationServiceClient;
 
 /** Delete Glossary */
-function sampleDeleteGlossary($projectId, $glossaryId)
-{
-    $translationServiceClient = new TranslationServiceClient();
+$translationServiceClient = new TranslationServiceClient();
 
-    // $projectId = '[Google Cloud Project ID]';
-    // $glossaryId = '[Glossary ID]';
-    $formattedName = $translationServiceClient->glossaryName($projectId, 'us-central1', $glossaryId);
+// $projectId = '[Google Cloud Project ID]';
+// $glossaryId = '[Glossary ID]';
+$formattedName = $translationServiceClient->glossaryName($projectId, 'us-central1', $glossaryId);
 
-    try {
-        $operationResponse = $translationServiceClient->deleteGlossary($formattedName);
-        $operationResponse->pollUntilComplete();
-        if ($operationResponse->operationSucceeded()) {
-            $response = $operationResponse->getResult();
-            printf('Deleted Glossary.' . PHP_EOL);
-        } else {
-            $error = $operationResponse->getError();
-            // handleError($error)
-        }
-    } finally {
-        $translationServiceClient->close();
+try {
+    $operationResponse = $translationServiceClient->deleteGlossary($formattedName);
+    $operationResponse->pollUntilComplete();
+    if ($operationResponse->operationSucceeded()) {
+        $response = $operationResponse->getResult();
+        printf('Deleted Glossary.' . PHP_EOL);
+    } else {
+        $error = $operationResponse->getError();
+        // handleError($error)
     }
+} finally {
+    $translationServiceClient->close();
 }
 // [END translate_v3_delete_glossary]
-
-$opts = [
-    'project_id::',
-    'glossary_id::',
-];
-
-$defaultOptions = [
-    'project_id' => '[Google Cloud Project ID]',
-    'glossary_id' => '[Glossary ID]',
-];
-
-$options = getopt('', $opts);
-$options += $defaultOptions;
-
-$projectId = $options['project_id'];
-$glossaryId = $options['glossary_id'];
-
-sampleDeleteGlossary($project, $glossaryId);
